Bernoulli pmf fails for OoB inputs; added docs
For int types, the Bernoulli PMF used to just check if the input was
equal to 1, in which case it would return p; otherwise it would return
(1 - p). So any value out of the support would cause the PMF to evaluate
to (1 - p) and not complain. That's bad. Now it checks that the input is
0, 1, or something else; and panics if it's something else.