Bump rand to 0.7
There are a lot of deprecation warnings about rand::distributions that
say to use the rand_distr crate instead. The problem is that all the
constructors in rand_distr do input validation, which we do not need or
want because we already do input validation. So we're going to deal with
the warnings for a bit.