Fix test macro; add macro for better cache cloning
In Clone implementations, the cache was cloned by creating a once cell
from the getter on the cached field. This would cause the cache to be
computed regardless of whether it was uninitialized. Now we have a
macro for f64 caches that calls `get` on the cache field, and only
copies the cached value if it has been computed.