Added some categorical prior tests that fail
Now that the `k` argument has been removed from the stat, and it
dynamically resizes its counts container, a couple of things will break
in ConjugatePrior when the stat is in a degenerate state. It used to be
that the stat could not be degenerate, but now the ln_m and ln_pp
functions will have be made to work with too-small count vectors and
panic with too-large count vectors.